Create Hazelcast node setup and initialization of cache

+++ src/main/java/ch/cern/c2mon/hazelcast/benchmarks/HazelcastGetBenchmark.java
@@ -15,7 +15,7 @@ public class HazelcastGetBenchmark extends AbstractHazelcastBenchmark {
 
   @Override
   public boolean test(Map<Object, Object> map) throws Exception {
-    int key = nextRandom(args.range());
+    long key = nextRandom(args.range());
 
     cache.get(key);
